Fixed android installation. Updated mainflow to handle lack of opengl context.

+++ bash/android/install_sdl.sh
@@ -8,14 +8,28 @@ sdl_install_directory="build/sdl/sdl"
 sdl_image_install_directory="build/sdl/sdl_image"
 sdl_mixer_install_directory="build/sdl/sdl_mixer"
 
-# Clone repositories
-git clone $SDL2_URL $sdl_install_directory
-git clone $SDL2_IMAGE_URL $sdl_image_install_directory
-git clone $SDL2_MIXER_URL $sdl_mixer_install_directory
+if [ -d $sdl_install_directory ]; then
+    echo "  > directory exists [$sdl_install_directory]"
+else
+    echo "  > directory does not exist [$sdl_install_directory]"
+    git clone -b release-2.26.x $SDL2_URL $sdl_install_directory
+    rm -rf $sdl_install_directory/.git
+fi
 
-# Clean up
-rm -rf $sdl_install_directory/.git
-rm -rf $sdl_image_install_directory/.git
-rm -rf $sdl_mixer_install_directory/.git
+if [ -d $sdl_image_install_directory ]; then
+    echo "  > directory exists [$sdl_image_install_directory]"
+else
+    echo "  > directory does not exist [$sdl_image_install_directory]"
+    git clone -b release-2.6.x $SDL2_IMAGE_URL $sdl_image_install_directory
+    rm -rf $sdl_image_install_directory/.git
+fi
 
-echo "SDL2, SDL2_image, and SDL2_mixer repositories downloaded successfully!"
+if [ -d $sdl_mixer_install_directory ]; then
+    echo "  > directory exists [$sdl_mixer_install_directory]"
+else
+    echo "  > directory does not exist [$sdl_mixer_install_directory]"
+    git clone -b release-2.6.x $SDL2_MIXER_URL $sdl_mixer_install_directory
+    rm -rf $sdl_mixer_install_directory/.git
+fi
+
+echo "  > sdl repositories downloaded successfully"
